+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-Better handled debug messages for anchors and muffled list
+Better handled debug messages for anchors and muffled list.
+Fixed config option to disable the inventory button not working.
+Added Muffler button to the creative inventory, with the corresponding config options.