+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-Better handled debug messages for anchors and muffled list.
-Fixed config option to disable the inventory button not working.
-Added Muffler button to the creative inventory, with the corresponding config options.
+Port to 1.19.4